move array related functions to `array_helpers`.
